Housing prices in Ames, Iowa

Data set contains information from the Ames Assessor's Office used in computing assessed values for individual residential properties sold in Ames, IA from 2006 to 2010. See http://www.amstat.org/publications/jse/v19n3/decock/datadocumentation.txt for detailed variable descriptions.

ames

Format

A tbl_df with with 2930 rows and 82 variables:

Order

Observation number.

PID

Parcel identification number - can be used with city web site for parcel review.

area

Above grade (ground) living area square feet.

price

Sale price in USD.

MS.SubClass

Identifies the type of dwelling involved in the sale.

MS.Zoning

Identifies the general zoning classification of the sale.

Lot.Frontage

Linear feet of street connected to property.

Lot.Area

Lot size in square feet.

Street

Type of road access to property.

Alley

Type of alley access to property.

Lot.Shape

General shape of property.

Land.Contour

Flatness of the property.

Utilities

Type of utilities available.

Lot.Config

Lot configuration.

Land.Slope

Slope of property.

Neighborhood

Physical locations within Ames city limits (map available).

Condition.1

Proximity to various conditions.

Condition.2

Proximity to various conditions (if more than one is present).

Bldg.Type

Type of dwelling.

House.Style

Style of dwelling.

Overall.Qual

Rates the overall material and finish of the house.

Overall.Cond

Rates the overall condition of the house.

Year.Built

Original construction date.

Year.Remod.Add

Remodel date (same as construction date if no remodeling or additions).

Roof.Style

Type of roof.

Roof.Matl

Roof material.

Exterior.1st

Exterior covering on house.

Exterior.2nd

Exterior covering on house (if more than one material).

Mas.Vnr.Type

Masonry veneer type.

Mas.Vnr.Area

Masonry veneer area in square feet.

Exter.Qual

Evaluates the quality of the material on the exterior.

Exter.Cond

Evaluates the present condition of the material on the exterior.

Foundation

Type of foundation.

Bsmt.Qual

Evaluates the height of the basement.

Bsmt.Cond

Evaluates the general condition of the basement.

Bsmt.Exposure

Refers to walkout or garden level walls.

BsmtFin.Type.1

Rating of basement finished area.

BsmtFin.SF.1

Type 1 finished square feet.

BsmtFin.Type.2

Rating of basement finished area (if multiple types).

BsmtFin.SF.2

Type 2 finished square feet.

Bsmt.Unf.SF

Unfinished square feet of basement area.

Total.Bsmt.SF

Total square feet of basement area.

Heating

Type of heating.

Heating.QC

Heating quality and condition.

Central.Air

Central air conditioning.

Electrical

Electrical system.

X1st.Flr.SF

First Floor square feet.

X2nd.Flr.SF

Second floor square feet.

Low.Qual.Fin.SF

Low quality finished square feet (all floors).

Bsmt.Full.Bath

Basement full bathrooms.

Bsmt.Half.Bath

Basement half bathrooms.

Full.Bath

Full bathrooms above grade.

Half.Bath

Half baths above grade.

Bedroom.AbvGr

Bedrooms above grade (does NOT include basement bedrooms).

Kitchen.AbvGr

Kitchens above grade.

Kitchen.Qual

Kitchen quality.

TotRms.AbvGrd

Total rooms above grade (does not include bathrooms).

Functional

Home functionality (Assume typical unless deductions are warranted).

Fireplaces

Number of fireplaces.

Fireplace.Qu

Fireplace quality.

Garage.Type

Garage location.

Garage.Yr.Blt

Year garage was built.

Garage.Finish

Interior finish of the garage.

Garage.Cars

Size of garage in car capacity.

Garage.Area

Size of garage in square feet.

Garage.Qual

Garage quality.

Garage.Cond

Garage condition.

Paved.Drive

Paved driveway.

Wood.Deck.SF

Wood deck area in square feet.

Open.Porch.SF

Open porch area in square feet.

Enclosed.Porch

Enclosed porch area in square feet.

X3Ssn.Porch

Three season porch area in square feet.

Screen.Porch

Screen porch area in square feet.

Pool.Area

Pool area in square feet.

Pool.QC

Pool quality.

Fence

Fence quality.

Misc.Feature

Miscellaneous feature not covered in other categories.

Misc.Val

Dollar value of miscellaneous feature.

Mo.Sold

Month Sold (MM).

Yr.Sold

Year Sold (YYYY).

Sale.Type

Type of sale.

Sale.Condition

Condition of sale.

Source

De Cock, Dean. "Ames, Iowa: Alternative to the Boston housing data as an end of semester regression project." Journal of Statistics Education 19.3 (2011).