Major League Baseball team data

Data from all 30 Major League Baseball teams from the 2011 season.

mlb11

Format

A data frame with 30 rows and 12 variables:

team

Team name.

runs

Number of runs.

at_bats

Number of at bats.

hits

Number of hits.

homeruns

Number of home runs.

bat_avg

Batting average.

strikeouts

Number of strikeouts.

stolen_bases

Number of stolen bases.

wins

Number of wins.

new_onbase

Newer variable: on-base percentage, a measure of how often a batter reaches base for any reason other than a fielding error, fielder's choice, dropped/uncaught third strike, fielder's obstruction, or catcher's interference.

new_slug

Newer variable: slugging percentage, popular measure of the power of a hitter calculated as the total bases divided by at bats.

new_obs

Newer variable: on-base plus slugging, calculated as the sum of the on-base and slugging percentages.

Source

mlb.com