Flights data

On-time data for a random sample of flights that departed NYC (i.e. JFK, LGA or EWR) in 2013.

nycflights

Format

A tbl_df with 32,735 rows and 16 variables:

year,month,day

Date of departure

dep_time,arr_time

Departure and arrival times, local tz.

dep_delay,arr_delay

Departure and arrival delays, in minutes. Negative times represent early departures/arrivals.

hour,minute

Time of departure broken in to hour and minutes

carrier

Two letter carrier abbreviation. See airlines in the nycflights13 package for more information

tailnum

Plane tail number

flight

Flight number

origin,dest

Origin and destination. See airports in the nycflights13 package for more information, or google airport the code.

air_time

Amount of time spent in the air

distance

Distance flown

Source

Hadley Wickham (2014). nycflights13: Data about flights departing NYC in 2013. R package version 0.1. https://CRAN.R-project.org/package=nycflights13